refact pagination

+++ src/Crosscutting/Infra.Crosscutting/Collections/IPagedList.cs
@@ -5,5 +5,5 @@ namespace Goal.Infra.Crosscutting.Collections;
 public interface IPagedList : IEnumerable
 {
     int PageLength { get; }
-    long TotalCount { get; }
+    int TotalCount { get; }
 }
